Implement cutoff insertion sort and refactor
Make separate static class 'Merge' for merging and class 'SortUtils' for checking that (sub)array is sorted. Also make test for all implementations and refactorings. Mergesort and version with cutoff are now two separate classes. Also some data has been made for the tests. For now they are just put in the unittest folder, but we might need more structure for input data soon. I can make a beforeeach in the test classes if deemed necessary. For now the arrays are manually read in the individual tests