implement initial adaptive version of lvlSort
Have made RunUtils class for finding runs in input. Have made tests for that class, and it should work pr specification. Right now it doesn't return compares, however. We have to decide whether we want it to, maybe in exchange for some overhead. Also have made test for the adaptive, and it passes. Still need to test it for edgecases specific to adaptiveness and for compares